sixel: Fix out-of-bounds write caused by bad resize logic
The finalization code shrinks the image extents if possible, but in
doing so it was only checking if one or the other dimension would be
reduced, leaving open the possibility that the other dimension could
be greater. When this happened, the result could exceed the buffer
provided by the caller.